+++ lib/shared/src/chat-models/index.ts
@@ -32,6 +32,13 @@ const defaultDotComChatModels: ChatModelProvider[] = [
         default: false,
         codyProOnly: true,
     },
+    {
+        title: 'Mixtral 8x7B',
+        model: 'fireworks/accounts/fireworks/models/mixtral-8x7b-instruct',
+        provider: 'Mistral',
+        default: false,
+        codyProOnly: true,
+    },
 ]
 
 /**